requirements.txt
python-dotenv==1.0.0
httpx==0.26.0
fastapi==0.115.8
langchain-community==0.3.18
langchain==0.3.19
langchain-core==0.3.39
openai==1.64.0
pydantic-settings==2.8.0
uvicorn==0.34.0
langchain-openai==0.3.7
langgraph>=0.2.56,<0.4.0
langgraph-sdk>=0.1.53
langgraph-checkpoint>=2.0.15,<3.0
langchain-google-genai==2.0.11
nest_asyncio==1.5.6
langchain-experimental>=0.3.4